Measuring analog frequency (side project)
UART frequency goes to audio out (L and R). There is a cable loopback and AD converter picks up audio L analog signal, and tries to measure the frequency. This kinda works:

for fout of 300, 19200 the picked up frequency is 2X
for 38400 doesn't work

The idea is to measure the amplitute of analog signal and flip a bit to 1 or 0 based on value. Min and max variable measure the amplitude for experimentation.