Tape record and load at 600bps 2 stop bits
Changed approach picking up the signal - the time between pulses (2400Hz for 0/space and 4800Hz for 1/mark) is measured using a high-frequency (0.75MHz) counter. R/S flip-flop detects is the time between pulses is below or above some time to regenerate the original signal. This is effectively a binary lo/hi pass filter. With this a 9k hex file has been recored to TI Program Recorder, tone dial was almost on Lo and volume dial around medium (which notch visible right in the middle). Tape was TDK PRO IEC1/TYPE1.